+++ rakaia/utils/cluster.py
@@ -95,7 +95,9 @@ def subset_cluster_frame(cluster_data: dict, roi_selection: str, clust_variable:
     """
     if cluster_data and roi_selection and clust_variable and cluster_cats and \
             roi_selection in cluster_data:
+        cluster_cats = [str(i) for i in cluster_cats]
         cluster_data = pd.DataFrame(cluster_data[roi_selection])
+        cluster_data[clust_variable] = cluster_data[clust_variable].apply(str)
         cluster_data = cluster_data[cluster_data[clust_variable].isin(list(cluster_cats))]
         if gating_object_list:
             object_column = get_cluster_proj_id_column(cluster_data)
@@ -103,26 +105,28 @@ def subset_cluster_frame(cluster_data: dict, roi_selection: str, clust_variable:
     return cluster_data
 
 def assign_colours_to_cluster_annotations(cluster_frame_dict: dict=None, cur_cluster_dict: dict=None,
-                                          roi_selection: str=None, cluster_id_col: str='cluster') -> tuple:
+                                          roi_selection: str=None) -> Union[dict, None]:
     """
     Generate a dictionary of random colours to assign to the clusters for a specific ROI
     cluster frame dict contains the cluster assignments by ROI
     cur_cluster_dict contains current assignments from previous uploads or previous ROIs
     """
     try:
-        unique_clusters = pd.DataFrame(cluster_frame_dict[roi_selection])[cluster_id_col].unique().tolist()
-        unique_colours = glasbey_palette(len(unique_clusters))
         cluster_assignments = {roi_selection: {}} if not cur_cluster_dict else cur_cluster_dict
         if roi_selection not in cluster_assignments:
             cluster_assignments[roi_selection] = {}
         cluster_assignments = match_cluster_hash_to_cluster_frame(cluster_frame_dict, cluster_assignments, roi_selection)
-        if cluster_id_col not in cluster_assignments[roi_selection].keys():
-            cluster_assignments[roi_selection][cluster_id_col] = {}
-            for clust, colour in zip(unique_clusters, unique_colours):
-                cluster_assignments[roi_selection][cluster_id_col][clust] = colour
-        return cluster_assignments, list(unique_clusters)
+        for cluster_cat in cluster_frame_dict[roi_selection].keys():
+            if cluster_cat not in ClusterIdentifiers.id_cols and \
+                cluster_cat not in cluster_assignments[roi_selection].keys():
+                cluster_assignments[roi_selection][cluster_cat] = {}
+                unique_clusters = pd.DataFrame(cluster_frame_dict[roi_selection])[cluster_cat].unique().tolist()
+                unique_colours = glasbey_palette(len(unique_clusters))
+                for clust, colour in zip(unique_clusters, unique_colours):
+                    cluster_assignments[roi_selection][cluster_cat][clust] = colour
+        return cluster_assignments
     except (KeyError, TypeError):
-        return None, None
+        return None
 
 def match_cluster_hash_to_cluster_frame(cluster_frame_dict: dict, cluster_assignments: dict,
                                         roi_selection: str) -> dict: